Split Ralph loop event handler
